Interpolation doesn't work

+++ Line.cs
@@ -11,7 +11,7 @@ namespace G__Marching_Sqaure
         public Point Start { get; set; }
         public Point End { get; set; }
 
-        public Line(int x1, int y1, int x2, int y2)
+        public Line(double x1, double y1, double x2, double y2)
         {
             Start = new Point(x1, y1);
             End = new Point(x2, y2);
